requirements.txt
wikipedia
faiss-cpu
python-dotenv
python-telegram-bot

# LangChain stack compatible con Google GenAI
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai==0.7.0
langchain-text-splitters
